viewer node callable name is on module as well
In other words, the callable can now be found on the module using
its name, not only using the 'main_callable' variable.

This is important because the feature of exporting node layouts
as python scripts rely on the name of the callables to find them
on the node scripts, not on the 'main_callable' variable.